requirements.txt
Flask
python-dotenv
requests
python-dateutil
gunicorn
numpy
